addScore api is created
