+++ api/src/main/java/com/community/api/endpoint/serviceProvider/ServiceProviderEntity.java
@@ -26,12 +26,13 @@ import java.util.Set;
 
 @Entity
 @Table(name = "service_provider")
-@Inheritance(strategy = InheritanceType.JOINED)
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
 @Setter
 @JsonIgnoreProperties({"hibernateLazyInitializer", "handler"})
+@DiscriminatorColumn(name = "type", discriminatorType = DiscriminatorType.STRING)
+@Inheritance(strategy = InheritanceType.JOINED)
 public class ServiceProviderEntity  {
 
     @Id
