+++ api/src/main/java/com/community/api/entity/ProfessionalServiceProvider.java
@@ -6,12 +6,14 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.Column;
+import javax.persistence.DiscriminatorValue;
 import javax.persistence.Entity;
 
 @Entity
 @Setter
 @Getter
 @NoArgsConstructor
+@DiscriminatorValue("PROFESSIONAL")
 public class ProfessionalServiceProvider extends ServiceProviderEntity
 {
     @Column(name = "business_unit_infra_score")
