+++ core/src/main/resources/META-INF/persistence-core.xml
@@ -61,7 +61,7 @@
 
         <exclude-unlisted-classes/>
     </persistence-unit>
-    
+
     <persistence-unit name="blSecurePU" transaction-type="RESOURCE_LOCAL">
         <non-jta-data-source>jdbc/webSecure</non-jta-data-source>
         <exclude-unlisted-classes/>
